Detect when Ekam is already running in the same directory.
If the second build is a oneshot build, just wait until the first build completes / becomes idle, since this means the build is in a state equivalent to what the oneshot build would have done (ignoring differing environments).

If the second build is a continuous build, error out, since we wouldn't want multiple continuous builds.

This change is useful particularly when you want to write a makefile that does a oneshot build as part of its process, but you also want to have a continuous build running for development.