Extend language server to automatically remove diagnostics when file is changed.
Currently, when the user modifies a file, vscode continues to show diagnostics at the same byte offset as they were originally, even though that no longer makes any sense.

If we wanted to be really fancy, we could subscribe to changes and actively update the locations of all diagnostics based on them. But, that's too much work. Instead, on any change, we just remove all diagnostics.