Correctly intercept `stat()` on glibc 2.33+.
glibc used to implement `stat()` as an inline wrapper in the header file, calling a symbol `__xstat()` which took a version number for backwards compatibility. So, we intercepted `__xstat()` but not `stat()`.

glibc 2.33 no longer does this. It exports `stat()` as a regular symbol. So we need to intercept it as such. We also still need to intercept `__xstat()` for binaries compiled with older versions of glibc.

Fixes #36.