Tell NM to use the matching LLVM-gold plugin
When building with LTO you need to use the LLVMGold plugin that matches
the version of Clang that's generating the LTO. Should be "safe" in
terms of not breaking any systems where this isn't available, but may
not enable LTO on all platforms (e.g. if the relative path between clang
and LLVMGold.so isn't "bin/../lib/LLVMGold.so".