Add EKAM_BYPASS_DIRS environment variable
This environment variable is a colon separated list of paths that have a
trailing '/' for which remapping should be skipped. This is useful for
things like ccache which try to access `~/.ccache` to write the cache
result. Since there's no actual guarantee that ccache is writing to
~/.ccache (could be overwritten as a setting or via runtime environment
variable), it's up to the end user to orchestrate this correctly.